backend/VietTuneArchive/Docker/requirements.txt
essentia-tensorflow==2.1.0
flask==3.0.0
numpy==1.24.0
librosa==0.10.0
scikit-learn==1.3.0
werkzeug==3.0.0
requests